task3

+++ src/main/java/com/ap/homebanking/models/Account.java
@@ -15,7 +15,7 @@ public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "native")
     @GenericGenerator(name = "native", strategy = "native")
-    private int id;
+    private long id;
     private String number;
     private LocalDate creationDate;
     private double balance;
@@ -35,11 +35,11 @@ public class Account {
         this.balance = balance;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-     public String getNumber() {
+    public String getNumber() {
         return number;
     }
